+++ internal/service/platform/usergroup/usergroup.go
@@ -52,12 +52,6 @@ func ResourceUserGroup() *schema.Resource {
 		Importer:      helpers.MultiLevelResourceImporter,
 
 		Schema: map[string]*schema.Schema{
-			// "is_sso_linked": {
-			// 	Description: "Whether the user group is linked to an SSO account.",
-			// 	Type:        schema.TypeBool,
-			// 	Optional:    true,
-			// 	Computed:    true,
-			// },
 			"linked_sso_id": {
 				Description: "The SSO account ID that the user group is linked to.",
 				Type:        schema.TypeString,
@@ -78,34 +72,40 @@ func ResourceUserGroup() *schema.Resource {
 			},
 			"notification_configs": {
 				Description: "List of notification settings.",
-				Type:        schema.TypeSet,
+				Type:        schema.TypeList,
 				Optional:    true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"type": {
-							Description: "Can be one of EMAIL, SLACK, PAGERDUTY, MSTEAMS",
+							Description: "Can be one of EMAIL, SLACK, PAGERDUTY, MSTEAMS.",
 							Type:        schema.TypeString,
 							Optional:    true,
 						},
 						"slack_webhook_url": {
-							Description: "Url of slack webhook",
+							Description: "Url of slack webhook.",
 							Type:        schema.TypeString,
 							Optional:    true,
 						},
 						"microsoft_teams_webhook_url": {
-							Description: "Url of Microsoft teams webhook",
+							Description: "Url of Microsoft teams webhook.",
 							Type:        schema.TypeString,
 							Optional:    true,
 						},
 						"pager_duty_key": {
-							Description: "Pager duty key",
+							Description: "Pager duty key.",
 							Type:        schema.TypeString,
 							Optional:    true,
 						},
 						"group_email": {
-							Description: "Group email",
+							Description: "Group email.",
 							Type:        schema.TypeString,
 							Optional:    true,
+						},
+						"send_email_to_all_users": {
+							Description: "Send email to all the group members.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Computed:    true,
 						},
 					},
 				},
@@ -126,12 +126,12 @@ func ResourceUserGroup() *schema.Resource {
 				Optional:    true,
 			},
 			"linked_sso_type": {
-				Description: "Type of linked SSO",
+				Description: "Type of linked SSO.",
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
 			"sso_linked": {
-				Description: "Whether sso is linked or not",
+				Description: "Whether sso is linked or not.",
 				Type:        schema.TypeBool,
 				Optional:    true,
 				Computed:    true,
@@ -249,7 +249,7 @@ func buildUserGroup(d *schema.ResourceData) nextgen.UserGroup {
 	}
 
 	if attr, ok := d.GetOk("notification_configs"); ok {
-		userGroup.NotificationConfigs = expandNotificationConfig(attr.(*schema.Set).List())
+		userGroup.NotificationConfigs = expandNotificationConfig(attr.([]interface{}))
 	}
 
 	if attr, ok := d.GetOk("is_sso_linked"); ok {
@@ -318,6 +318,7 @@ func expandNotificationConfig(notificationConfigs []interface{}) []nextgen.Notif
 		}
 		if resultNotificationConfig.Type_ == "EMAIL" {
 			resultNotificationConfig.GroupEmail = v["group_email"].(string)
+			resultNotificationConfig.SendEmailToAllUsers = v["send_email_to_all_users"].(bool)
 		}
 		if resultNotificationConfig.Type_ == "MSTEAMS" {
 			resultNotificationConfig.MicrosoftTeamsWebhookUrl = v["microsoft_teams_webhook_url"].(string)
@@ -341,8 +342,9 @@ func flattenNotificationConfig(notificationConfigs []nextgen.NotificationSetting
 		}
 		if notificationConfig.Type_ == "EMAIL" {
 			result = append(result, map[string]interface{}{
-				"type":        notificationConfig.Type_,
-				"group_email": notificationConfig.GroupEmail,
+				"type":                    notificationConfig.Type_,
+				"group_email":             notificationConfig.GroupEmail,
+				"send_email_to_all_users": notificationConfig.SendEmailToAllUsers,
 			})
 		}
 		if notificationConfig.Type_ == "MSTEAMS" {